Transform: Align pointers to 16-byte boundaries
'ovec' is a contiguous collection of many different pointers. Each of
these may not be aligned, depending on the sizes that came before. We
round each sub-alignment to 16 bytes, so that the resulting pointers are
always on 16 byte boundaries.

On certain configurations (e.g. Ubuntu's armhf builders), unaligned
access raises a SIGBUS which terminates the program, so this is a real
problem.

Fixes #200.